Added ffnbot!refresh.

+++ ffn_bot/reddit_bot.py
@@ -22,14 +22,15 @@ DEFAULT_SUBREDDITS = ['HPFanfiction']
 SUBREDDIT_LIST = set()
 CHECKED_COMMENTS = None
 FOOTER = "\n".join([
-    r"**Bot v1.2.0 - 9/6/15** **|** \[[Usage][1]\] | \[[Changelog][2]\] | \[[Issues][3]\] | \[[GitHub][4]\]",
+    r"**Bot v1.3.0 - 9/7/15** **|** \[[Usage][1]\] | \[[Changelog][2]\] | \[[Issues][3]\] | \[[GitHub][4]\]",
     r'[1]: https://github.com/tusing/reddit-ffn-bot/wiki/Usage       "How to use the bot"',
     r'[2]: https://github.com/tusing/reddit-ffn-bot/wiki/Changelog   "What changed until now"',
     r'[3]: https://github.com/tusing/reddit-ffn-bot/issues/          "Bugs? Suggestions? Enter them here!"',
     r'[4]: https://github.com/tusing/reddit-ffn-bot/                 "Fork me on GitHub"'
 ])
 FOOTER += "\n\n^^^^^^^^^^^^^^^^^ffnbot!ignore"
-FOOTER += "\n\n**Update Notes:** *Use *ffnbot!delete* to delete a comment!*"
+FOOTER += "\n\n**Update Notes:** *Use* **ffnbot!delete** *to delete a comment!"
+FOOTER +=  "Use* **ffnbot!refresh** *to refresh a bot replies!*"
 
 # For testing purposes
 DRY_RUN = False
@@ -197,14 +198,15 @@ def handle_submission(submission, markers=frozenset()):
 def handle_comment(comment, extra_markers=frozenset()):
     logging.debug("Handling comment: " + comment.id)
     if (str(comment.id) not in CHECKED_COMMENTS
-        ) or ("force" in extra_markers):
+            ) or ("force" in extra_markers):
 
-        logging.info("Found new comment: " + comment.id)
         markers = parse_context_markers(comment.body)
         markers |= extra_markers
         if "ignore" in markers:
-            logging.info("Comment forcefully ignored: " + comment.id)
+            # logging.info("Comment forcefully ignored: " + comment.id)
             return
+        else:
+            logging.info("Found new comment: " + comment.id)
 
         if "parent" in markers:
             if comment.is_root:
@@ -213,7 +215,8 @@ def handle_comment(comment, extra_markers=frozenset()):
                 item = r.get_info(thing_id=comment.parent_id)
             handle(item, {"directlinks", "submissionlink", "force"})
 
-        if "delete" in markers:
+        if "delete" in markers and (comment.id not in CHECKED_COMMENTS):
+            CHECKED_COMMENTS.add(str(comment.id))
             logging.info("Delete requested by " + comment.id)
             if not (comment.is_root):
                 parent_comment = r.get_info(thing_id=comment.parent_id)
@@ -228,20 +231,63 @@ def handle_comment(comment, extra_markers=frozenset()):
             else:
                 logging.error("Delete requested by invalid comment!")
 
-        if "refresh" in markers:
-            logging.info("Refresh requested by " + comment.id)
-            comment_with_requests = r.get_info(thing_id=comment.parent_id)
-            if comment_with_requests.author is None:
-                logging.info(
-                    "(Refresh) Original comment with requests is invalid.")
-                return
-            reply_list = comment_with_requests.replies
+        if "refresh" in markers and (str(comment.id) not in CHECKED_COMMENTS):
+            CHECKED_COMMENTS.add(str(comment.id))
+            logging.info("(Refresh) Refresh requested by " + comment.id)
 
-            if reply_list is not None:
+            # Get the full comment or submission
+            comment_with_requests = get_full(comment.parent_id)
+            logging.info("(Refresh) Refreshing on " + type(
+                comment_with_requests).__name__ + " with id " + comment_with_requests.id)
+
+            # TODO: Make it so FanfictionBot does not have to be hardcoded
+            # If ffnbot!refresh is called on an actual bot reply, then go up
+            # one level to find the requesting comment
+            if comment_with_requests.author.name == "FanfictionBot":
+                logging.info(
+                    "(Refresh) Refresh requested on a bot comment (" + comment_with_requests.id + ").")
+                # Retrieve the requesting parent submission or comment
+                comment_with_requests = get_full(
+                    comment_with_requests.parent_id)
+
+                # If the requesting comment has been deleted, abort
+                if not valid_comment(comment_with_requests):
+                    logging.error(
+                        "(Refresh) Parent of bot comment is invalid.")
+                    return
+
+                logging.info(
+                    "          Refresh request being pushed to parent " + comment_with_requests.id)
+
+            if isinstance(comment_with_requests, praw.objects.Comment):
+                logging.info(
+                    "(Refresh) Running refresh on COMMENT " + str(comment_with_requests.id))
+                logging.info("(Refresh) Appending replies to deletion check list: " +
+                             ", ".join(str(c.id) for c in comment_with_requests.replies))
+                delete_list = comment_with_requests.replies
+
+            elif isinstance(comment_with_requests, praw.objects.Submission):
+                logging.info(
+                    "(Refresh) Running refresh on SUBMISSION " + str(comment_with_requests.id))
+
+                unfiltered_delete_list = comment_with_requests.comments
+                delete_list = []
+                for comment in unfiltered_delete_list:
+                    if comment.author is not None:
+                        if (comment.author.name == "FanfictionBot"):
+                            delete_list.append(comment)
+                            print("(Refresh) Found root-level bot comment " + comment.id)
+            else:
+                logging.error("(Refresh) Can't refresh " + comment_with_requests.type(
+                ).__name__ + " with ID " + comment_with_requests.id)
+                bot_tools.pause(5, 0)
+                return
+
+            if delete_list is not None:
                 logging.info("(Refresh) Finding replies to delete.")
-                for reply in reply_list:
-                    if reply.author is not None:
-                        if (reply.author == "FanfictionBot"):
+                for reply in delete_list:
+                    if valid_comment(reply):
+                        if (reply.author.name == "FanfictionBot"):
                             logging.error(
                                 "(Refresh) Deleting bot comment " + reply.id)
                             reply.delete()
@@ -249,13 +295,52 @@ def handle_comment(comment, extra_markers=frozenset()):
                 logging.info(
                     "(Refresh) No bot replies have been made. Continuing...")
             CHECKED_COMMENTS.add(str(comment.id))
-            handle_comment(comment_with_requests, frozenset(["force"]))
+
+            if isinstance(comment_with_requests, praw.objects.Comment):
+                logging.info(
+                    "(Refresh) Re-handling comment " + comment_with_requests.id)
+                handle_comment(comment_with_requests, frozenset(["force"]))
+            elif isinstance(comment_with_requests, praw.objects.Submission):
+                logging.info(
+                    "(Refresh) Re-handling submission " + comment_with_requests.id)
+                handle_submission(comment_with_requests, frozenset(["force"]))
             return
 
         try:
             make_reply(comment.body, comment.id, comment.reply, markers)
         finally:
             CHECKED_COMMENTS.add(str(comment.id))
+
+
+def get_full(comment_id):
+    """
+    Will return a full comment or submission.
+    Very heavy on time.
+    """
+    requested_comment = r.get_info(thing_id=comment_id)
+    if isinstance(requested_comment, praw.objects.Comment):
+        # PRAW doesn't return replies in a comment object retrieved with
+        # get_info; we must do this:
+        requested_comment = r.get_submission(
+            requested_comment.permalink).comments[0]
+    elif isinstance(requested_comment, praw.objects.Submission):
+        requested_comment = r.get_submission(requested_comment.permalink)
+        requested_comment.replace_more_comments(limit=None, threshold=0)
+    else:
+        logging.error(
+            "(URGENT) WAS NOT ABLE TO DETERMINE COMMENT VS SUBMISSION!")
+        requested_comment = r.get_submission(requested_comment.permalink)
+    return requested_comment
+
+
+def valid_comment(comment):
+    """
+    Checks if valid comment.
+    """
+    if comment.author is None:
+        logging.error("Found invalid comment " + comment.id)
+        return False
+    return True
 
 
 def handle(obj, markers=frozenset()):
